Fixing unit tests to work with the new legacy tree and moved classes.
