--- tests/suite/cms/toolbar/JToolbarButtonTest.php
@@ -1,61 +1,0 @@
-<?php
-/**
- * @package     Joomla.UnitTest
- * @subpackage  HTML
- *
- * @copyright   Copyright (C) 2005 - 2012 Open Source Matters, Inc. All rights reserved.
- * @license     GNU General Public License version 2 or later; see LICENSE
- */
-
-require_once JPATH_PLATFORM . '/cms//toolbar/button.php';
-
-/**
- * Test class for JButton.
- * Generated by PHPUnit on 2009-10-27 at 16:59:36.
- */
-class JToolbarButtonTest extends PHPUnit_Framework_TestCase
-{
-	/**
-	 * @todo Implement testGetName().
-	 */
-	public function testGetName()
-	{
-		// Remove the following lines when you implement this test.
-		$this->markTestIncomplete(
-		'This test has not been implemented yet.'
-		);
-	}
-
-	/**
-	 * @todo Implement testRender().
-	 */
-	public function testRender()
-	{
-		// Remove the following lines when you implement this test.
-		$this->markTestIncomplete(
-		'This test has not been implemented yet.'
-		);
-	}
-
-	/**
-	 * @todo Implement testFetchIconClass().
-	 */
-	public function testFetchIconClass()
-	{
-		// Remove the following lines when you implement this test.
-		$this->markTestIncomplete(
-		'This test has not been implemented yet.'
-		);
-	}
-
-	/**
-	 * @todo Implement testFetchId().
-	 */
-	public function testFetchId()
-	{
-		// Remove the following lines when you implement this test.
-		$this->markTestIncomplete(
-		'This test has not been implemented yet.'
-		);
-	}
-}
