--- tests/suite/cms/toolbar/button/JButtonHelpTest.php
@@ -1,40 +1,0 @@
-<?php
-/**
- * @package     Joomla.UnitTest
- * @subpackage  HTML
- *
- * @copyright   Copyright (C) 2005 - 2012 Open Source Matters, Inc. All rights reserved.
- * @license     GNU General Public License version 2 or later; see LICENSE
- */
-
-require_once JPATH_PLATFORM . '/cms/toolbar/button.php';
-require_once JPATH_PLATFORM . '/cms/toolbar/button/help.php';
-
-/**
- * Test class for JButtonHelp.
- * Generated by PHPUnit on 2009-10-27 at 17:01:55.
- */
-class JToolbarButtonHelpTest extends PHPUnit_Framework_TestCase
-{
-	/**
-	 * @todo Implement testFetchButton().
-	 */
-	public function testFetchButton()
-	{
-		// Remove the following lines when you implement this test.
-		$this->markTestIncomplete(
-		'This test has not been implemented yet.'
-		);
-	}
-
-	/**
-	 * @todo Implement testFetchId().
-	 */
-	public function testFetchId()
-	{
-		// Remove the following lines when you implement this test.
-		$this->markTestIncomplete(
-		'This test has not been implemented yet.'
-		);
-	}
-}
