--- tests/suite/joomla/application/JCategoriesTest.php
@@ -1,35 +1,0 @@
-<?php
-/**
- * @package     Joomla.UnitTest
- * @subpackage  Application
- *
- * @copyright   Copyright (C) 2005 - 2012 Open Source Matters, Inc. All rights reserved.
- * @license     GNU General Public License version 2 or later; see LICENSE
- */
-
-require_once JPATH_PLATFORM.'/joomla/application/categories.php';
-
-/**
- * Test class for JCategories.
- * Generated by PHPUnit on 2009-10-08 at 11:55:02.
- */
-class JCategoriesTest extends PHPUnit_Framework_TestCase
-{
-	/**
-	 * @todo Implement testGetInstance().
-	 */
-	public function testGetInstance()
-	{
-		// Remove the following lines when you implement this test.
-		$this->markTestIncomplete('This test has not been implemented yet.');
-	}
-
-	/**
-	 * @todo Implement testGet().
-	 */
-	public function testGet()
-	{
-		// Remove the following lines when you implement this test.
-		$this->markTestIncomplete('This test has not been implemented yet.');
-	}
-}
