--- tests/suite/joomla/application/JCategoryNodeTest.php
@@ -1,62 +1,0 @@
-<?php
-/**
- * @package     Joomla.UnitTest
- * @subpackage  Application
- *
- * @copyright   Copyright (C) 2005 - 2012 Open Source Matters, Inc. All rights reserved.
- * @license     GNU General Public License version 2 or later; see LICENSE
- */
-
-require_once JPATH_PLATFORM.'/joomla/application/categories.php';
-
-/**
- * Test class for JCategoryNode.
- * Generated by PHPUnit on 2009-10-08 at 11:55:26.
- */
-class JCategoryNodeTest extends PHPUnit_Framework_TestCase
-{
-	/**
-	 * @todo Implement testAddChild().
-	 */
-	public function testAddChild()
-	{
-		// Remove the following lines when you implement this test.
-		$this->markTestIncomplete('This test has not been implemented yet.');
-	}
-
-	/**
-	 * @todo Implement testGetParent().
-	 */
-	public function testGetParent()
-	{
-		// Remove the following lines when you implement this test.
-		$this->markTestIncomplete('This test has not been implemented yet.');
-	}
-
-	/**
-	 * @todo Implement testSetParent().
-	 */
-	public function testSetParent()
-	{
-		// Remove the following lines when you implement this test.
-		$this->markTestIncomplete('This test has not been implemented yet.');
-	}
-
-	/**
-	 * @todo Implement testHasChildren().
-	 */
-	public function testHasChildren()
-	{
-		// Remove the following lines when you implement this test.
-		$this->markTestIncomplete('This test has not been implemented yet.');
-	}
-
-	/**
-	 * @todo Implement testGetChildren().
-	 */
-	public function testGetChildren()
-	{
-		// Remove the following lines when you implement this test.
-		$this->markTestIncomplete('This test has not been implemented yet.');
-	}
-}
