--- tests/suite/joomla/error/JProfilerTest.php
@@ -1,90 +1,0 @@
-<?php
-/**
- * @package     Joomla.UnitTest
- *
- * @copyright   Copyright (C) 2005 - 2012 Open Source Matters, Inc. All rights reserved.
- * @license     GNU General Public License version 2 or later; see LICENSE
- */
-
-include_once JPATH_PLATFORM . '/joomla/error/profiler.php';
-
-/**
- * Test class for JProfiler.
- * Generated by PHPUnit on 2011-10-26 at 19:34:10.
- */
-class JProfilerTest extends PHPUnit_Framework_TestCase {
-
-    /**
-     * @var JProfiler
-     */
-    protected $object;
-
-    /**
-     * Sets up the fixture, for example, opens a network connection.
-     * This method is called before a test is executed.
-     */
-    protected function setUp() {
-        $this->object = new JProfiler;
-    }
-
-    /**
-     * Tears down the fixture, for example, closes a network connection.
-     * This method is called after a test is executed.
-     */
-    protected function tearDown() {
-
-    }
-
-    /**
-     * @todo Implement testGetInstance().
-     */
-    public function testGetInstance() {
-        // Remove the following lines when you implement this test.
-        $this->markTestIncomplete(
-                'This test has not been implemented yet.'
-        );
-    }
-
-    /**
-     * @todo Implement testMark().
-     */
-    public function testMark() {
-        // Remove the following lines when you implement this test.
-        $this->markTestIncomplete(
-                'This test has not been implemented yet.'
-        );
-    }
-
-    /**
-     * @todo Implement testGetmicrotime().
-     */
-    public function testGetmicrotime() {
-        // Remove the following lines when you implement this test.
-        $this->markTestIncomplete(
-                'This test has not been implemented yet.'
-        );
-    }
-
-    /**
-     * @todo Implement testGetMemory().
-     */
-    public function testGetMemory() {
-        // Remove the following lines when you implement this test.
-        $this->markTestIncomplete(
-                'This test has not been implemented yet.'
-        );
-    }
-
-    /**
-     * @todo Implement testGetBuffer().
-     */
-    public function testGetBuffer() {
-        // Remove the following lines when you implement this test.
-        $this->markTestIncomplete(
-                'This test has not been implemented yet.'
-        );
-    }
-
-}
-
-?>
