+++ tests/suite/legacy/request/JRequestTest.php
@@ -1,0 +1,148 @@
+<?php
+/**
+ * @package     Joomla.UnitTest
+ * @subpackage  Environment
+ *
+ * @copyright   Copyright (C) 2005 - 2012 Open Source Matters, Inc. All rights reserved.
+ * @license     GNU General Public License version 2 or later; see LICENSE
+ */
+
+require_once __DIR__ . '/JRequest-helper-dataset.php';
+
+/**
+ * Test class for JRequest.
+ * Generated by PHPUnit on 2009-10-09 at 14:02:01.
+ */
+class JRequestTest extends PHPUnit_Framework_TestCase {
+	/**
+	 * @var	JRequest
+	 */
+	protected $object;
+
+	/**
+	 * Sets up the fixture, for example, opens a network connection.
+	 * This method is called before a test is executed.
+	 */
+	protected function setUp() {
+
+	}
+
+	/**
+	 * Tears down the fixture, for example, closes a network connection.
+	 * This method is called after a test is executed.
+	 */
+	protected function tearDown() {
+	}
+
+	/**
+	 * @todo Implement testGetURI().
+	 */
+	public function testGetURI() {
+		// Remove the following lines when you implement this test.
+		$this->markTestIncomplete('This test has not been implemented yet.');
+	}
+
+	/**
+	 * @todo Implement testGetMethod().
+	 */
+	public function testGetMethod() {
+		// Remove the following lines when you implement this test.
+		$this->markTestIncomplete('This test has not been implemented yet.');
+	}
+
+	/**
+	 * @todo Implement testGetVar().
+	 */
+	public function testGetVar() {
+		// Remove the following lines when you implement this test.
+		$this->markTestIncomplete('This test has not been implemented yet.');
+	}
+
+	/**
+	 * @todo Implement testGetInt().
+	 */
+	public function testGetInt() {
+		// Remove the following lines when you implement this test.
+		$this->markTestIncomplete('This test has not been implemented yet.');
+	}
+
+	/**
+	 * @todo Implement testGetFloat().
+	 */
+	public function testGetFloat() {
+		// Remove the following lines when you implement this test.
+		$this->markTestIncomplete('This test has not been implemented yet.');
+	}
+
+	/**
+	 * @todo Implement testGetBool().
+	 */
+	public function testGetBool() {
+		// Remove the following lines when you implement this test.
+		$this->markTestIncomplete('This test has not been implemented yet.');
+	}
+
+	/**
+	 * @todo Implement testGetWord().
+	 */
+	public function testGetWord() {
+		// Remove the following lines when you implement this test.
+		$this->markTestIncomplete('This test has not been implemented yet.');
+	}
+
+	/**
+	 * @todo Implement testGetCmd().
+	 */
+	public function testGetCmd() {
+		// Remove the following lines when you implement this test.
+		$this->markTestIncomplete('This test has not been implemented yet.');
+	}
+
+	/**
+	 * @todo Implement testGetString().
+	 */
+	public function testGetString() {
+		// Remove the following lines when you implement this test.
+		$this->markTestIncomplete('This test has not been implemented yet.');
+	}
+
+	/**
+	 * @todo Implement testSetVar().
+	 */
+	public function testSetVar() {
+		// Remove the following lines when you implement this test.
+		$this->markTestIncomplete('This test has not been implemented yet.');
+	}
+
+	/**
+	 * @todo Implement testGet().
+	 */
+	public function testGet() {
+		// Remove the following lines when you implement this test.
+		$this->markTestIncomplete('This test has not been implemented yet.');
+	}
+
+	/**
+	 * @todo Implement testSet().
+	 */
+	public function testSet() {
+		// Remove the following lines when you implement this test.
+		$this->markTestIncomplete('This test has not been implemented yet.');
+	}
+
+	/**
+	 * @todo Implement testCheckToken().
+	 */
+	public function testCheckToken() {
+		// Remove the following lines when you implement this test.
+		$this->markTestIncomplete('This test has not been implemented yet.');
+	}
+
+	/**
+	 * @todo Implement testClean().
+	 */
+	public function testClean() {
+		// Remove the following lines when you implement this test.
+		$this->markTestIncomplete('This test has not been implemented yet.');
+	}
+}
