+++ tests/suite/legacy/toolbar/button/JButtonSeparatorTest.php
@@ -1,0 +1,37 @@
+<?php
+/**
+ * @package     Joomla.UnitTest
+ * @subpackage  HTML
+ *
+ * @copyright   Copyright (C) 2005 - 2012 Open Source Matters, Inc. All rights reserved.
+ * @license     GNU General Public License version 2 or later; see LICENSE
+ */
+
+/**
+ * Test class for JButtonSeparator.
+ * Generated by PHPUnit on 2009-10-27 at 17:04:11.
+ */
+class JToolbarButtonSeparatorTest extends PHPUnit_Framework_TestCase
+{
+	/**
+	 * @todo Implement testRender().
+	 */
+	public function testRender()
+	{
+		// Remove the following lines when you implement this test.
+		$this->markTestIncomplete(
+		'This test has not been implemented yet.'
+		);
+	}
+
+	/**
+	 * @todo Implement testFetchButton().
+	 */
+	public function testFetchButton()
+	{
+		// Remove the following lines when you implement this test.
+		$this->markTestIncomplete(
+		'This test has not been implemented yet.'
+		);
+	}
+}
